backend/requirements.txt
fastapi
uvicorn
send2trash
duplicate_images
pillow
